moved common functions to common.ts and updated unit tests
